+++ src/mefit/Forms/frmSettings.cs
@@ -245,8 +245,8 @@ namespace Mac_EFI_Toolkit.Forms
 
         private void SetButtonProperties()
         {
-            cmdClose.Font = Program.SegoeFluentRegular12;
-            cmdClose.Text = ApplicationChars.FLUENT_MULTIPLY;
+            cmdClose.Font = Program.FluentRegular12;
+            cmdClose.Text = ApplicationChars.FLUENT_DISMISS;
         }
         #endregion
     }