+++ src/mefit/Updater.cs
@@ -6,7 +6,6 @@
 
 using Mac_EFI_Toolkit.Common;
 using Mac_EFI_Toolkit.Common.Constants;
-using Mac_EFI_Toolkit.Tools;
 using System;
 using System.Diagnostics;
 using System.IO;
@@ -60,7 +59,7 @@ namespace Mac_EFI_Toolkit
 
                         NewVersion = xmlVersionNode.InnerText;
                         ExpectedSHA256 = xmlSha256Node.InnerText;
-                        Priority = xmlPriorityNode.InnerText;
+                        // Priority = xmlPriorityNode.InnerText;
 
                         Console.WriteLine($"{nameof(CheckForNewVersion)} -> {nameof(NewVersion)} '{xmlVersionNode.InnerText}'");
                         Console.WriteLine($"{nameof(CheckForNewVersion)} -> {nameof(ExpectedSHA256)} '{xmlSha256Node.InnerText}'");